+++ packages/contracts/contracts/Migrations.sol
@@ -4,7 +4,7 @@ pragma solidity 0.4.24;
 contract Migrations {
 
     address public owner;
-    uint public lastCompletedMigration;
+    uint256 public lastCompletedMigration;
 
     modifier restricted() {
         if (msg.sender == owner) _;
@@ -14,7 +14,7 @@ contract Migrations {
         owner = msg.sender;
     }
 
-    function setCompleted(uint _completed) public restricted {
+    function setCompleted(uint256 _completed) public restricted {
         lastCompletedMigration = _completed;
     }
